Added dif.java and changed the variable type in sum.java from int to double

+++ sum.java
@@ -1,8 +1,8 @@
 public class SumTwoNumbers {
     public static void main(String[] args) {
-        int num1 = 10; // First number
-        int num2 = 20; // Second number
-        int sum = num1 + num2; // Calculate the sum
+        double num1 = 10; // First number
+        double num2 = 20; // Second number
+        double sum = num1 + num2; // Calculate the sum
 
         System.out.println("The sum of " + num1 + " and " + num2 + " is: " + sum);
     }
